requirements.txt
Django==5.2.3
djangorestframework==3.16.0
drf-yasg==1.21.10
python-decouple==3.8
hashids==1.3.1
validators==0.35.0
django-cors-headers==4.7.0